message shown when you reach 140 characters started on
