+++ code/src/components/Input.css
@@ -30,3 +30,9 @@ label {
   border: none;
   padding: 8px;
 }
+
+.counter {
+  font-size: 10px;
+  margin: 0px 0px 10px 0px;
+  font-weight: 100;
+}
